Move functionality to a dedicated component
